Crate gateway_tests

Structs

GatewayTestOpts ğŸ”’

Enums

GatewayTest ğŸ”’

Functions

backup_restore_test ğŸ”’

check_empty_lightning_balance ğŸ”’

check_empty_onchain_balance ğŸ”’

config_test ğŸ”’

Test that sets and verifies configurations within the gateway

esplora_test ğŸ”’

This test cannot be run in CI because it connects to an external esplora server

get_transaction ğŸ”’

leave_federation ğŸ”’

Leaves the specified federation by issuing a leave-fed POST request to the gateway.

liquidity_test ğŸ”’

Test that verifies the various liquidity tools (onchain, lightning, ecash) work correctly.

main ğŸ”’

mnemonic_upgrade_test ğŸ”’

TODO(v0.5.0): We do not need to run the gatewayd-mnemonic test from v0.4.0 -> v0.5.0 over and over again. Once we have verified this test passes for v0.5.0, it can safely be removed.

stop_and_recover_gateway ğŸ”’