Crate gateway_tests

Structs

GatewayTestOpts ğŸ”’

Enums

GatewayTest ğŸ”’

Functions

backup_restore_test ğŸ”’

config_test ğŸ”’

Test that sets and verifies configurations within the gateway

get_gateway_info ğŸ”’

Retrieves the GatewayInfo by issuing an info GET request to the gateway.

leave_federation ğŸ”’

Leaves the specified federation by issuing a leave-fed POST request to the gateway.

main ğŸ”’

mnemonic_upgrade_test ğŸ”’

TODO(v0.5.0): We do not need to run the gatewayd-mnemonic test from v0.4.0 -> v0.5.0 over and over again. Once we have verified this test passes for v0.5.0, it can safely be removed.

stop_and_recover_gateway ğŸ”’